More printf fixes
git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@36414 379a1393-f5fb-40a0-bcee-ef074d9b53f7

+++ vice/src/arch/gtk3/widgets/digimaxwidget.c
@@ -67,15 +67,15 @@ static void on_digimax_toggled(GtkWidget *widget, gpointer user_data)
  */
 static void on_combo_changed(GtkWidget *widget, gpointer user_data)
 {
-    int value;
+    long value;
     char *endptr;
     const char *id_str;
 
     id_str = gtk_combo_box_get_active_id(GTK_COMBO_BOX(widget));
-    value = (int)strtol(id_str, &endptr, 10);
+    value = strtol(id_str, &endptr, 10);
     if (*endptr == '\0') {
-        debug_gtk3("setting DIGIMAXbase to $%04X\n", value);
-        resources_set_int("DIGIMAXbase", value);
+        debug_gtk3("setting DIGIMAXbase to $%04X\n", (unsigned int)value);
+        resources_set_int("DIGIMAXbase", (int)value);
     }
 }
